Simulator fixes and support for using MacOS: * Added simulator support for Mac. * Fix for simulator to properly assemble wolfboot.bin + signedtestapp + update + swap. * Fixes for handling 64-bit assigned mmap virtual addresses. Added hal_flash_write and hal_flash_erase support for 64-bit address using uintptr_t. Enabled if platform is 64-bit and `FORCE_32BIT` is not defined * Fix simulator conflict with src/libwolfboot.o object in test-app. * Cleanup test-app linker flags.

+++ hal/raspi3.c
@@ -114,7 +114,7 @@ void hal_prepare_boot(void)
 }
 
 
-int RAMFUNCTION hal_flash_write(uint32_t address, const uint8_t *data, int len)
+int RAMFUNCTION hal_flash_write(uintptr_t address, const uint8_t *data, int len)
 {
     return 0;
 }
@@ -128,7 +128,7 @@ void RAMFUNCTION hal_flash_lock(void)
 }
 
 
-int RAMFUNCTION hal_flash_erase(uint32_t address, int len)
+int RAMFUNCTION hal_flash_erase(uintptr_t address, int len)
 {
     return 0;
 }
